Enable compilation with CLANG: * replace "extern inline" by "static inline" * add some void casts to shut up warnings * use printf("%s", foo) instead of printf(foo) * don't use -traditional * use {} for empty if statements instead of ; * fix an asm constraint
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@2041 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/base/mouse/mouse.c
@@ -1639,7 +1639,7 @@ static void call_int15_mouse_event_handler(void)
     mouse.old_mickeyx += dx;
     mouse.old_mickeyy -= dy;
 
-    m_printf("PS2MOUSE data: dx=%hhx, dy=%hhx, status=%hx\n", (Bit8u)dx, (Bit8u)dy, status);
+    m_printf("PS2MOUSE data: dx=%hhx, dy=%hhx, status=%hx\n", (Bit8u)dx, (Bit8u)dy, (unsigned short)status);
     pushw(ssp, sp, status | 8);
     pushw(ssp, sp, dx & 0xff);
     pushw(ssp, sp, dy & 0xff);
